Rename githubChecks to verifyCommitStatusChecks
